Struct bls_dkg::key_gen::KeyGen

pub struct KeyGen { /* fields omitted */ }

An algorithm for dealerless distributed key generation.

This is trying to follow the protocol as suggested at https://github.com/dashpay/dips/blob/master/dip-0006/bls_m-of-n_threshold_scheme_and_dkg.md#distributed-key-generation-dkg-protocol

A normal usage flow will be: a, call initialize first to generate an instance. b, multicasting the return Message to all participants. c, call handle_message function to handle the incoming Message and multicasting the resulted Message (if has) to all participants. d, call finalize_complaining_phase to complete the complaining phase. (This separate call may need to depend on a separate timer & checker against the key generator’s current status) e, repeat step c when there is incoming Message. f, call generate_keys to get the public-key set and secret-key share, if the procedure finalized.

Implementations

impl KeyGen

pub fn initialize(
our_id: XorName,
threshold: usize,
names: BTreeSet<XorName>
) -> Result<(KeyGen, Message), Error>

Creates a new KeyGen instance, together with the Initial message that should be multicast to all nodes.

pub fn phase(&self) -> Phase

pub fn handle_message<R: RngCore>(
&mut self,
rng: &mut R,
msg: Message
) -> Result<Vec<Message>, Error>

Dispatching an incoming dkg message.

pub fn all_contribution_received(&self) -> bool

pub fn timed_phase_transition<R: RngCore>(
&mut self,
rng: &mut R
) -> Result<Vec<Message>, Error>

pub fn is_finalized(&self) -> bool

Returns true if in the phase of Finalization.

pub fn generate_keys(&self) -> Option<(BTreeSet<XorName>, Outcome)>

Returns the new secret key share and the public key set.

pub fn possible_blockers(&self) -> BTreeSet<XorName>

This function shall be called when the DKG procedure not reach Finalization phase and before discarding the instace. It returns potential invalid peers that causing the blocking, if any and provable.