Module bls_dkg::key_gen

Modules

message

outcome

Structs

Acknowledgment
A confirmation that we have received and verified a validator's part. It must be sent to all participating nodes and handled by all of them, including ourselves.

KeyGen
An algorithm for dealerless distributed key generation.

Part
A contribution by a node for the key generation. The part shall only be handled by the receiver.

Enums

AcknowledgmentFault
Acknowledgment faulty entries.

Error
A local error while handling a message, that was not caused by that message being invalid.

PartFault
Part faulty entries.

PartOutcome
The outcome of handling and verifying a Part message.

Phase